base/rag/indexers/domain/requirements.txt
httpx>=0.24.0
pymilvus>=2.4.0
pyyaml>=6.0
